Release v2.0.0.
This version drops compatibility for Node.js <= 10.x, and as such is a
semver major version bump.

To be clear, the code in this package hasn't changed; we've just
bumped the versions of dependencies.  The `commander` dep requires
node 6 ( https://github.com/tj/commander.js/pull/1053 ) but the dev
dependencies require Node.js <= 8.x and are unsupported on Node.js <=
10.x.  I can't promise that things that aren't being tested continue
to work.